+++ heylogs-cli/src/main/java/nbbrd/heylogs/cli/CheckCommand.java
@@ -1,12 +1,13 @@
 package nbbrd.heylogs.cli;
 
 import internal.heylogs.cli.MarkdownInputOptions;
-import nbbrd.heylogs.Failure;
-import nbbrd.heylogs.Rule;
+import nbbrd.console.picocli.text.TextOutputOptions;
+import nbbrd.heylogs.*;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
-import java.nio.file.Path;
+import java.io.Writer;
+import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.Callable;
 
@@ -16,31 +17,41 @@ public final class CheckCommand implements Callable<Integer> {
     @CommandLine.Mixin
     private MarkdownInputOptions input;
 
+    @CommandLine.Mixin
+    private TextOutputOptions output;
+
+    @CommandLine.Option(
+            names = {"-f", "--format"},
+            defaultValue = "stylish",
+            description = "Specify the formatter used to control the appearance of the result. Valid values: ${COMPLETION-CANDIDATES}.",
+            completionCandidates = FormatCandidates.class
+    )
+    private String format;
+
     @Override
     public Integer call() throws Exception {
-        List<Rule> rules = Rule.getDefault();
+        List<Failure> failures = Failure.allOf(input.read(), RuleLoader.load());
 
-        List<Failure> failures = Failure.allOf(input.read(), rules);
-
-        printStylish(input.getFile(), failures);
+        try (Writer writer = output.newCharWriter()) {
+            FailureFormatterLoader.load()
+                    .stream()
+                    .filter(formatter -> formatter.getName().equals(format))
+                    .findFirst()
+                    .orElse(new StylishFormatter())
+                    .format(writer, input.hasFile() ? input.getFile().toString() : "stdin", failures);
+        }
 
         return failures.isEmpty() ? CommandLine.ExitCode.OK : CommandLine.ExitCode.USAGE;
     }
 
-    // https://eslint.org/docs/latest/user-guide/formatters/#stylish
-    private static void printStylish(Path inputFile, List<Failure> failures) {
-        if (inputFile != null) {
-            System.out.println(inputFile);
+    public static final class FormatCandidates implements Iterable<String> {
+
+        @Override
+        public Iterator<String> iterator() {
+            return FailureFormatterLoader.load()
+                    .stream()
+                    .map(FailureFormatter::getName)
+                    .iterator();
         }
-
-        int l = failures.stream().mapToInt(failure -> getNumberOfDigits(failure.getLine())).max().orElse(0);
-        int c = failures.stream().mapToInt(failure -> getNumberOfDigits(failure.getColumn())).max().orElse(0);
-        int m = failures.stream().mapToInt(failure -> failure.getMessage().length()).max().orElse(0);
-
-        failures.forEach(failure -> System.out.println(String.format("  %" + l + "d:%-" + c + "d  error  %-" + m + "s  %s", failure.getLine(), failure.getColumn(), failure.getMessage(), failure.getRule())));
-    }
-
-    private static int getNumberOfDigits(int number) {
-        return (int) (Math.log10(number) + 1);
     }
 }