+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/CheckMojo.java
@@ -3,7 +3,7 @@ package nbbrd.heylogs.maven.plugin;
 import com.vladsch.flexmark.parser.Parser;
 import com.vladsch.flexmark.util.ast.Document;
 import nbbrd.heylogs.Failure;
-import nbbrd.heylogs.Rule;
+import nbbrd.heylogs.RuleLoader;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
@@ -19,7 +19,7 @@ import java.util.List;
 @Mojo(name = "check", defaultPhase = LifecyclePhase.GENERATE_RESOURCES, threadSafe = true)
 public final class CheckMojo extends AbstractMojo {
 
-    @Parameter(defaultValue = "${project.basedir}/CHANGELOG.md", property = "heylogs.inputFile")
+    @Parameter(defaultValue = "${project.basedir}/CHANGELOG.md", property = "heylogs.input.file")
     private File inputFile;
 
     @Parameter(defaultValue = "false", property = "heylogs.skip")
@@ -41,7 +41,7 @@ public final class CheckMojo extends AbstractMojo {
             getLog().info("Reading " + inputFile);
             Document changelog = read();
 
-            List<Failure> failures = Failure.allOf(changelog, Rule.getDefault());
+            List<Failure> failures = Failure.allOf(changelog, RuleLoader.load());
             if (!failures.isEmpty()) {
                 getLog().error("Invalid CHANGELOG");
                 failures.forEach(failure -> getLog().error(failure.toString()));