+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [0.3.0] - 2022-09-27
+
+### Added
+
+- Add rule extension point [#12](https://github.com/nbbrd/heylogs/issues/12)
+- Add failure format extension point [#13](https://github.com/nbbrd/heylogs/issues/13)
+- Add unreleased pattern to version filtering [#19](https://github.com/nbbrd/heylogs/issues/19)
+
+### Changed
+
+- Simplify properties naming of Maven plugin [#18](https://github.com/nbbrd/heylogs/issues/18)
+
 ## [0.2.0] - 2022-09-21
 
 ### Added
@@ -17,6 +31,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/nbbrd/heylogs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/nbbrd/heylogs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/nbbrd/heylogs/releases/tag/v0.1.0